pyproject.toml
[tool.poetry]
name = "meguchi"
version = "0.1.0"
description = "Ogura Megumi"
authors = ["nothink <gakongakon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.3"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.5.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
flake8 = "^3.8"
black = {version = "*", allow-prereleases = true}
flake8-black = "^0.2.1"
pytest-django = "^4.1.0"
factory-boy = "^3.1.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
